Track allocator looks for long index. Shift is now properly in the +Z axis

+++ src/Thread_allocation/ThreadAllocator.cc
@@ -143,7 +143,7 @@ void ThreadAllocator::MakeTasks(const TString& file, std::vector<ThreadTask>& in
 	//Establish a pointer to and readout the end point of the tracks
 	long start_entry_val = 0;
 	TTreeReader trackIndexReader("Track index", &f);
-	TTreeReaderValue<long long> end_entry_val(trackIndexReader, "index");
+	TTreeReaderValue<long> end_entry_val(trackIndexReader, "index");
 
 	//Make the output vector
 	std::vector<ThreadTask> output;
